% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binomialize}
\alias{binomialize}
\title{binomialize}
\usage{
binomialize(df)
}
\arguments{
\item{df:}{dataframe containing risky trials}
}
\value{
dataframe
}
\description{
format df into binomial version for model fitting
}
