% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ipsi_bias}
\alias{compute_ipsi_bias}
\title{compute_ipsi_bias}
\usage{
compute_ipsi_bias(choice, side)
}
\arguments{
\item{choice:}{a vector with 0s and 1s}

\item{side:}{'left', 'right'}
}
\value{
dataframe
}
\description{
computes the ipsilateral side bias
helper function for figure_learning
}
