% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeELPD}
\alias{computeELPD}
\title{computeELPD}
\usage{
computeELPD(log_lik_heldout)
}
\arguments{
\item{log_lik_heldout:}{}
}
\value{
dataframe
}
\description{
adapted from https://github.com/stan-dev/stancon_talks/blob/master/2017/Contributed-Talks/07_nicenboim/kfold.Rmd
the ELPD is the theoretical expected log pointwise predictive density for the held-out data
the loglik from the fit is a n_iter x T matrix, the pointwise is 1 x T vector, elpd is the sum of pointwise
}
