% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_example}
\alias{figure_behavior_example}
\title{figure_behavior_example}
\usage{
figure_behavior_example()
}
\value{
ggplot object
}
\description{
figure1c: example animal psychometric curves by GLMM, thin lines for each session, thick line for combined
}
