% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_population}
\alias{figure_behavior_population}
\title{figure_behavior_population}
\usage{
figure_behavior_population()
}
\value{
ggplot object
}
\description{
figure1d: population psychometric curves by GLMM, thin colored lines per subject, thick line for population
}
