% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_session_stat}
\alias{figure_behavior_session_stat}
\title{figure_behavior_session_stat}
\usage{
figure_behavior_session_stat()
}
\value{
ggplot object
}
\description{
figure1e: average number of trials in session vs. number of sessions, colored by subject
}
