% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_behavior.R
\name{figure_behavior_timeline}
\alias{figure_behavior_timeline}
\title{figure_behavior_timeline}
\usage{
figure_behavior_timeline(n = 30)
}
\arguments{
\item{n:}{the number of trials to show}
}
\value{
ggplot object
}
\description{
figure1b: timeline of trials in one example session.
}
