% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_control.R
\name{figure_control_bar}
\alias{figure_control_bar}
\title{figure_control_bar}
\usage{
figure_control_bar()
}
\value{
ggplot object
}
\description{
figure3c: bar plot showing mean absolute change in % ipsilateral bias in free choice and risky choices
}
