% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_control.R
\name{figure_control_bias}
\alias{figure_control_bias}
\title{figure_control_bias}
\usage{
figure_control_bias()
}
\value{
ggplot object
}
\description{
figure3b: compare ipsiateral bias before, during and after infusoin for both free choices and risky choices
}
