% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_control.R
\name{figure_control_risky}
\alias{figure_control_risky}
\title{figure_control_risky}
\usage{
figure_control_risky()
}
\value{
ggplot object
}
\description{
figure3d: plots psychometric curves (model prediction + binned data), color by unilateral infusion and control
}
