% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_infusion.R
\name{figure_infusion_bi}
\alias{figure_infusion_bi}
\title{figure_infusion_bi}
\usage{
figure_infusion_bi(region = "FOF")
}
\arguments{
\item{region:}{'FOF' or 'PPC'}
}
\value{
ggplot object
}
\description{
figure2b & 2e: plots psychometric curves (model prediction + binned data), color by bilateral infusion and control
}
