% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_infusion.R
\name{figure_infusion_side}
\alias{figure_infusion_side}
\title{figure_infusion_side}
\usage{
figure_infusion_side()
}
\value{
ggplot object
}
\description{
figure2d: plots psychometric curves (model prediction + binned data), color by one side infusion and control
}
