% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_infusion.R
\name{figure_infusion_uni}
\alias{figure_infusion_uni}
\title{figure_infusion_uni}
\usage{
figure_infusion_uni(region = "FOF")
}
\arguments{
\item{region}{= 'FOF' or 'PPC'}
}
\value{
ggplot object
}
\description{
figure2c & 2f: plots psychometric curves (model prediction + binned data), color by unilateral infusion and control
}
