% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_learning.R
\name{figure_learning_scatter}
\alias{figure_learning_scatter}
\title{figure_learning_scatter}
\usage{
figure_learning_scatter()
}
\value{
ggplot object
}
\description{
figure7d: scatter plot of model predicted change vs. actual change in probability choosing lottery, color by infusion and control
}
