% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_learning.R
\name{figure_learning_schematic}
\alias{figure_learning_schematic}
\title{figure_learning_schematic}
\usage{
figure_learning_schematic()
}
\value{
ggplot object
}
\description{
figure7a: schematic showing how shifting the surebet magnitude shifts the decision boundary
}
