% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_density}
\alias{figure_model_density}
\title{figure_model_density}
\usage{
figure_model_density()
}
\value{
ggplot object
}
\description{
figure4d: density plot for each parameter using concatenated posteriors from all subjects
}
