% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_prediction}
\alias{figure_model_prediction}
\title{figure_model_prediction}
\usage{
figure_model_prediction()
}
\value{
ggplot object
}
\description{
figure4b & figureS8b: plots psychometric curves (model ribbon prediction + binned data)
If fitting a new model, the ribbon might be slightly different each time
}
