% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_model.R
\name{figure_model_utility}
\alias{figure_model_utility}
\title{figure_model_utility}
\usage{
figure_model_utility()
}
\value{
ggplot object
}
\description{
figure4c: use rho estimates from the model to plot utility curves for each subject
}
