% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_shift.R
\name{figure_shift_elpd}
\alias{figure_shift_elpd}
\title{figure_shift_elpd}
\usage{
figure_shift_elpd()
}
\value{
ggplot object
}
\description{
figure5c: compare ELPD from 10-fold cross-validation results using the rho-only, omega-only and rho-and-omega inactivation model
}
