% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_shift.R
\name{figure_shift_individual_deviation}
\alias{figure_shift_individual_deviation}
\title{figure_shift_individual_deviation}
\usage{
figure_shift_individual_deviation(model = "rho-and-omega")
}
\arguments{
\item{model}{= 'rho-and-omega', 'rho-only', 'omega-only'}
}
\value{
ggplot object
}
\description{
figure5b: plot individual posterior deviation of FOF inactivation to control (matrix, row = rat, column = parameter)
}
