% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_shift.R
\name{figure_shift_prediction}
\alias{figure_shift_prediction}
\title{figure_shift_prediction}
\usage{
figure_shift_prediction()
}
\value{
ggplot object
}
\description{
figure5a & figureS10: model prediction ribbons overlyed with control and FOF-inactivation binned choice data
}
