% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_control_pairs}
\alias{figureS_control_pairs}
\title{figureS_control_pairs}
\usage{
figureS_control_pairs()
}
\value{
ggplot object
}
\description{
figureS9: aggregated pairs plot with control posterior distributions
}
