% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_individual_posterior}
\alias{figureS_individual_posterior}
\title{figureS_individual_posterior}
\usage{
figureS_individual_posterior(model = "rho-and-omega")
}
\arguments{
\item{model:}{'rho-and-omega', 'rho-only', 'omega-only' for different inactivation models}
}
\value{
ggplot object
}
\description{
figureS11: plot individual posterior distribution in a matrix, where row = subjid, column = parameter, dataset = color
}
