% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_infusion}
\alias{figureS_infusion}
\title{figureS_infusion}
\usage{
figureS_infusion(region = "bi_fof")
}
\arguments{
\item{region:}{'bi_fof', 'bi_ppc', 'uni_fof', 'uni_ppc'}
}
\value{
ggplot object
}
\description{
figureS4 & S5: GLMM fits and plots for Bilateral FOF/PPC, Unilateral FOF/PPC
}
