% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_RT}
\alias{figureS_RT}
\title{figureS_RT}
\usage{
figureS_RT(region = "bi_fof")
}
\arguments{
\item{region:}{'bi_fof', 'bi_ppc', 'uni_fof', 'uni_ppc'}
}
\value{
ggplot object
}
\description{
figureS6 & S7: LMM fits and plots for reaction time in Bilateral FOF/PPC, Unilateral FOF/PPC
}
