% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_sanity}
\alias{figureS_sanity}
\title{figureS_sanity}
\usage{
figureS_sanity(N = 20)
}
\arguments{
\item{N:}{number of synthetic datasets}
}
\value{
ggplot object
}
\description{
figure S8a: generate model parameters from the prior range, simulate dataset, and see if the model can recover them accurately
returns a scatter plot of predictions vs. true values
the plot will be different each time it's run
}
