% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS.R
\name{figureS_timeline}
\alias{figureS_timeline}
\title{figureS_timeline}
\usage{
figureS_timeline(animal = 2156)
}
\arguments{
\item{animal:}{a subjid in the list of infusion_animals}
}
\value{
ggplot object
}
\description{
figureS1: timeline plot of all the experimental manipulations performed on an animal
}
