% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_animals}
\alias{fit_animals}
\title{fit_animals}
\usage{
fit_animals(dataset = "control")
}
\arguments{
\item{dataset:}{'control', 'bi_fof', 'bi_ppc}
}
\value{
dataframe
}
\description{
fit all animal data with the individual model and save fit details in csv/
}
