% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_animals_inactivation}
\alias{fit_animals_inactivation}
\title{fit_animals_inactivation}
\usage{
fit_animals_inactivation(fit_model, model = "rho-only")
}
\arguments{
\item{fit_model:}{compiled stan model corresponding to the inactivation model used}

\item{model:}{'rho-and-omega', 'rho-only', 'omega-only'}
}
\value{
dataframe
}
\description{
fit all animal data with the inactivation model and save in csv/
}
