% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_syn_smooth}
\alias{gen_syn_smooth}
\title{gen_syn_smooth}
\usage{
gen_syn_smooth(df, sb_mags = NULL, extend = "delta_ev")
}
\arguments{
\item{df:}{dataframe with risky trials}

\item{sb_mags:}{a list with sb_mags to include, if NULL it will get the most frequent sb_mag}

\item{extend:}{extend by 'lottery_mag' or 'delta_ev'}
}
\value{
dataframe
}
\description{
generate smooth synthetic data points for every task parameter combinations
}
