% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_freq_task_param}
\alias{get_freq_task_param}
\title{get_freq_task_param}
\usage{
get_freq_task_param(df, param = "sb_mag")
}
\arguments{
\item{df:}{dataframe with risky trials}

\item{param:}{task parameter name, 'sb_mag', 'lottery_mag', 'lottery_prob' and 'total_rew_multi'}
}
\value{
numeric
}
\description{
find the most frequent task parameter given the df
}
