% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_control.R
\name{plot_before_infusion_after}
\alias{plot_before_infusion_after}
\title{plot_before_infusion_after}
\usage{
plot_before_infusion_after(plot_df)
}
\value{
ggplot object
}
\description{
helper function for figure_control_compare_bias
}
