% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_risk}
\alias{plot_risk}
\title{plot_risk}
\usage{
plot_risk(df, subjid = 0, by_prev_outcome = FALSE)
}
\arguments{
\item{df:}{dataframe containing risky trials}

\item{subjid:}{subjid}

\item{by_prev_outcome:}{Boolean, if true it plots separately for lottery-win, lottery-lose and sure-bet as previous outcome}
}
\value{
ggplot object
}
\description{
plots binned risk choices with specified subjid
}
