% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho_sigma_agent}
\alias{rho_sigma_agent}
\title{rho_sigma_agent}
\usage{
rho_sigma_agent(
  prob = FALSE,
  params,
  sb_mag,
  lottery_mag,
  lottery_prob,
  total_rew_multi
)
}
\arguments{
\item{prob:}{Boolean, if true returns choice_prob}

\item{params:}{a list as in list('rho' = 1, 'sigma' = 0.4, 'omega' = c(0.8, 0.1, 0.1))}

\item{sb_mag:}{a vector of n_trials with surebet magnitude}

\item{lottery_mag:}{a vector of n_trials with lottery magnitude}

\item{lottery_prob:}{a vector of n_trials with lottery probability}

\item{total_rew_multi:}{a vector of n_trials with total reward multiplier}
}
\value{
a vector of n_trials with choice_prob or choice
}
\description{
returns choice_prob (probability choosing lottery) or choice (binary) from the simulated three-agent rho-sigma agent
}
