% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_save}
\alias{scale_save}
\title{scale_save}
\usage{
scale_save(p, name, width = 4, height = 4, scale = 1)
}
\arguments{
\item{p:}{a ggplot object to be saved}

\item{name:}{the name of the file}

\item{width:}{the width of the plot}

\item{height:}{the height of the plot}

\item{scale:}{scale of the plot, 1 is the default}
}
\value{
saves the plot in the specified path
}
\description{
ggplot wrapper to save the plot
}
